Struct google_run1::api::NamespaceMethods

pub struct NamespaceMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on namespace resources. It is not used directly, but through the CloudRun hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_run1 as run1;
 
use std::default::Default;
use oauth2;
use run1::CloudRun;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = CloudRun::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `authorizeddomains_list(...)`, `configurations_get(...)`, `configurations_list(...)`, `domainmappings_create(...)`, `domainmappings_delete(...)`, `domainmappings_get(...)`, `domainmappings_list(...)`, `revisions_delete(...)`, `revisions_get(...)`, `revisions_list(...)`, `routes_get(...)`, `routes_list(...)`, `services_create(...)`, `services_delete(...)`, `services_get(...)`, `services_list(...)` and `services_replace_service(...)`
// to build up your call.
let rb = hub.namespaces();

Implementations

impl<'a, C> NamespaceMethods<'a, C>

pub fn authorizeddomains_list(
&self,
parent: &str
) -> NamespaceAuthorizeddomainListCall<'a, C>

Create a builder to help you perform the following task:

List authorized domains.

Arguments

parent - Name of the parent Project resource. Example: projects/myproject.

pub fn configurations_get(
&self,
name: &str
) -> NamespaceConfigurationGetCall<'a, C>

Create a builder to help you perform the following task:

Get information about a configuration.

Arguments

name - The name of the configuration to retrieve. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.

pub fn configurations_list(
&self,
parent: &str
) -> NamespaceConfigurationListCall<'a, C>

Create a builder to help you perform the following task:

List configurations.

Arguments

parent - The namespace from which the configurations should be listed. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.

pub fn domainmappings_create(
&self,
request: DomainMapping,
parent: &str
) -> NamespaceDomainmappingCreateCall<'a, C>

Create a builder to help you perform the following task:

Create a new domain mapping.

Arguments

request - No description provided.

parent - The namespace in which the domain mapping should be created. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.

pub fn domainmappings_delete(
&self,
name: &str
) -> NamespaceDomainmappingDeleteCall<'a, C>

Create a builder to help you perform the following task:

Delete a domain mapping.

Arguments

name - The name of the domain mapping to delete. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.

pub fn domainmappings_get(
&self,
name: &str
) -> NamespaceDomainmappingGetCall<'a, C>

Create a builder to help you perform the following task:

Get information about a domain mapping.

Arguments

name - The name of the domain mapping to retrieve. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.

pub fn domainmappings_list(
&self,
parent: &str
) -> NamespaceDomainmappingListCall<'a, C>

Create a builder to help you perform the following task:

List domain mappings.

Arguments

parent - The namespace from which the domain mappings should be listed. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.

pub fn revisions_delete(&self, name: &str) -> NamespaceRevisionDeleteCall<'a, C>

Create a builder to help you perform the following task:

Delete a revision.

Arguments

name - The name of the revision to delete. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.

pub fn revisions_get(&self, name: &str) -> NamespaceRevisionGetCall<'a, C>

Create a builder to help you perform the following task:

Get information about a revision.

Arguments

name - The name of the revision to retrieve. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.

pub fn revisions_list(&self, parent: &str) -> NamespaceRevisionListCall<'a, C>

Create a builder to help you perform the following task:

List revisions.

Arguments

parent - The namespace from which the revisions should be listed. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.

pub fn routes_get(&self, name: &str) -> NamespaceRouteGetCall<'a, C>

Create a builder to help you perform the following task:

Get information about a route.

Arguments

name - The name of the route to retrieve. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.

pub fn routes_list(&self, parent: &str) -> NamespaceRouteListCall<'a, C>

Create a builder to help you perform the following task:

List routes.

Arguments

parent - The namespace from which the routes should be listed. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.

pub fn services_create(
&self,
request: Service,
parent: &str
) -> NamespaceServiceCreateCall<'a, C>

Create a builder to help you perform the following task:

Create a service.

Arguments

request - No description provided.

parent - The namespace in which the service should be created. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.

pub fn services_delete(&self, name: &str) -> NamespaceServiceDeleteCall<'a, C>

Create a builder to help you perform the following task:

Delete a service. This will cause the Service to stop serving traffic and will delete the child entities like Routes, Configurations and Revisions.

Arguments

name - The name of the service to delete. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.

pub fn services_get(&self, name: &str) -> NamespaceServiceGetCall<'a, C>

Create a builder to help you perform the following task:

Get information about a service.

Arguments

name - The name of the service to retrieve. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.

pub fn services_list(&self, parent: &str) -> NamespaceServiceListCall<'a, C>

Create a builder to help you perform the following task:

List services.

Arguments

parent - The namespace from which the services should be listed. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.

pub fn services_replace_service(
&self,
request: Service,
name: &str
) -> NamespaceServiceReplaceServiceCall<'a, C>

Create a builder to help you perform the following task:

Replace a service. Only the spec and metadata labels and annotations are modifiable. After the Update request, Cloud Run will work to make the 'status' match the requested 'spec'. May provide metadata.resourceVersion to enforce update from last read for optimistic concurrency control.

Arguments

request - No description provided.

name - The name of the service being replaced. For Cloud Run (fully managed), replace {namespace_id} with the project ID or number.